Automatically load tbox into ontojournal for GPAD export command.

+++ minerva-core/src/main/java/org/geneontology/minerva/BlazegraphOntologyManager.java
@@ -66,7 +66,7 @@ public class BlazegraphOntologyManager {
 	private final static String public_blazegraph_url = "http://skyhook.berkeleybop.org/blazegraph-go-lego-reacto-neo.jnl.gz";
 	//TODO this should probably go somewhere else - like an ontology file - this was missing..
 	public static String in_taxon_uri = "https://w3id.org/biolink/vocab/in_taxon";
-	public static OWLAnnotationProperty in_taxon;
+	public static OWLAnnotationProperty in_taxon = OWLManager.getOWLDataFactory().getOWLAnnotationProperty(IRI.create(in_taxon_uri));
 	private static final Set<String> root_types;
 	public final Map<IRI, Set<IRI>> regulatorsToRegulated;
 	public Map<String, Integer> class_depth;
@@ -89,24 +89,20 @@ public class BlazegraphOntologyManager {
 		root_types.add("http://purl.obolibrary.org/obo/ECO_0000000"); //evidence root.  
 	}
 
-	public BlazegraphOntologyManager(String go_lego_repo_file) throws IOException {
-		OWLOntologyManager ontman = OWLManager.createOWLOntologyManager();	
-		in_taxon = ontman.getOWLDataFactory().getOWLAnnotationProperty(IRI.create(in_taxon_uri));
-		if(new File(go_lego_repo_file).exists()) {			
-			go_lego_repo = initializeRepository(go_lego_repo_file);
-		}else {
-			LOG.info("No blazegraph tbox journal found at "+go_lego_repo_file+" . Downloading from "+public_blazegraph_url+" and putting there.");
+	public BlazegraphOntologyManager(String go_lego_repo_file, boolean downloadJournal) throws IOException {
+		if (!new File(go_lego_repo_file).exists() && downloadJournal) {
+			LOG.info("No blazegraph tbox journal found at " + go_lego_repo_file + " . Downloading from " + public_blazegraph_url + " and putting there.");
 			URL blazegraph_url = new URL(public_blazegraph_url);
 			File go_lego_repo_local = new File(go_lego_repo_file);
-			if(public_blazegraph_url.endsWith(".gz")) {
-				go_lego_repo_local = new File(go_lego_repo_file+".gz");
+			if (public_blazegraph_url.endsWith(".gz")) {
+				go_lego_repo_local = new File(go_lego_repo_file + ".gz");
 			}
 			org.apache.commons.io.FileUtils.copyURLToFile(blazegraph_url, go_lego_repo_local);
-			if(public_blazegraph_url.endsWith(".gz")) {
-				unGunzipFile(go_lego_repo_file+".gz", go_lego_repo_file);
+			if (public_blazegraph_url.endsWith(".gz")) {
+				unGunzipFile(go_lego_repo_file + ".gz", go_lego_repo_file);
 			}
-			go_lego_repo = initializeRepository(go_lego_repo_file);
 		}
+		go_lego_repo = initializeRepository(go_lego_repo_file);
 		class_depth = buildClassDepthMap("http://purl.obolibrary.org/obo/GO_0003674");
 		class_depth.putAll(buildClassDepthMap("http://purl.obolibrary.org/obo/GO_0008150"));
 		class_depth.putAll(buildClassDepthMap("http://purl.obolibrary.org/obo/GO_0005575"));
@@ -190,8 +186,32 @@ public class BlazegraphOntologyManager {
 			} finally {
 				connection.close();
 			}
-			return ;
-		}		
+		}
+	}
+
+	public void loadRepositoryFromOntology(OWLOntology ontology, String iri, boolean reset) throws OWLOntologyCreationException, RepositoryException, IOException, RDFParseException, RDFHandlerException {
+		synchronized(go_lego_repo) {
+			final BigdataSailRepositoryConnection connection = go_lego_repo.getUnisolatedConnection();
+			try {
+				connection.begin();
+				try {
+					URI graph = new URIImpl(iri);
+					if(reset) {
+						connection.clear(graph);
+					}
+					StatementCollector collector = new StatementCollector();
+					RioRenderer renderer = new RioRenderer(ontology, collector, null);
+					renderer.render();
+					connection.add(collector.getStatements(), graph);
+					connection.commit();
+				} catch (Exception e) {
+					connection.rollback();
+					throw e;
+				}
+			} finally {
+				connection.close();
+			}
+		}
 	}
 
 	public Set<String> getAllSuperClasses(String uri) throws IOException {
